vesnin: Add SRESET signal support
VESNIN server uses NMI SRESET signal to initiate memory dumping with
kdump in case of crash of the host OS. It is not necessary to return
control to the host's kernel since it has already crashed. This
commit allows for registration and handling of
opal_signal_system_reset signal for POWER8 based systems.

Resolves SRV-526

Signed-off-by: Maxim Polyakov <m.polyakov@yadro.com>