vesnin: remove #address/#size-cell in bmc dt node
The child node ("sensors") from “bmc” in the device tree doesn`t
contain "reg" property with address information. For this reason,
the properties #address-cell and #size-cell isn`t needed in "bmc"
node. The patch for the hostboot removes these properties.

In addition, it allows the dtc utility to get rid of the warning:

(avoid_unnecessary_addr_size): /bmc: unnecessary #address-cells
/#size-cells without "ranges" or child "reg" property

https://github.com/open-power/hostboot/pull/170

Signed-off-by: Maxim Polyakov <m.polyakov@yadro.com>